+++ src/shared/Route/ProtectedRoute.tsx
@@ -1,11 +1,11 @@
 import { Navigate } from 'react-router-dom';
-import LoginAlertModal from '@/pages/Login/LoginAlertModal';
+import Modal from '@/shared/components/organisms/Modal';
 
 export default function ProtectedRoute({ children }: { children: JSX.Element }) {
   const isLoggedIn = localStorage.getItem('isLoggedIn') === 'true';
 
   if (!isLoggedIn) {
-    LoginAlertModal();
+    Modal({ icon: 'error', title: '로그인이 필요합니다.', buttonTitle: '확인' });
     return <Navigate to="/" replace />;
   }
