+++ src/shared/components/atoms/SnsButton.tsx
@@ -1,7 +1,7 @@
 import { isMobile } from 'react-device-detect';
 import { IoChatbubble } from 'react-icons/io5';
 import { Images } from '@/shared/assets/images';
-import FeaturePendingModal from '@/shared/components/organisms/FeaturePendingModal';
+import Modal from '@/shared/components/organisms/Modal';
 
 const SnsButton = ({ type }: { type: 'kakao' | 'github' }) => {
   return (
@@ -9,7 +9,7 @@ const SnsButton = ({ type }: { type: 'kakao' | 'github' }) => {
       {type === 'kakao' ? (
         <div
           className={`flex-center ${isMobile ? 'w-full' : 'w-[27.375rem]'} h-[2.75rem] bg-[#FEE500] text-black font-bold rounded-[0.25rem] cursor-pointer`}
-          onClick={FeaturePendingModal}
+          onClick={() => Modal({ icon: 'warning', desc: '준비중인 기능입니다.', buttonTitle: '확인' })}
         >
           <IoChatbubble className="h-5 w-5 mr-1 mb-1" />
           <button>카카오로 시작하기</button>
@@ -18,7 +18,7 @@ const SnsButton = ({ type }: { type: 'kakao' | 'github' }) => {
       {type === 'github' ? (
         <div
           className={`flex-center  ${isMobile ? 'w-full' : 'w-[27.375rem]'} h-[2.75rem] bg-white text-black font-bold rounded-[0.25rem] border border-disabled cursor-pointer`}
-          onClick={FeaturePendingModal}
+          onClick={() => Modal({ icon: 'warning', desc: '준비중인 기능입니다.', buttonTitle: '확인' })}
         >
           <img src={Images.GithubLogo} className="h-5 mr-1 mb-1" />
           <button>깃허브로 시작하기</button>
